+++ ext/Server/Bot/BotAttacking.lua
@@ -56,8 +56,7 @@ function BotAttacking:UpdateAttacking(p_Bot)
 
 			if p_Bot._ActiveAction == BotActionFlags.GrenadeActive then -- throw grenade
 				if p_Bot.m_Player.soldier.weaponsComponent.currentWeapon.secondaryAmmo <= 0 then
-					p_Bot.m_Player.soldier.weaponsComponent.currentWeapon.secondaryAmmo = p_Bot.m_Player.soldier.weaponsComponent.currentWeapon
-						.secondaryAmmo + 1
+					p_Bot.m_Player.soldier.weaponsComponent.currentWeapon.secondaryAmmo = 1
 					p_Bot:_ResetActionFlag(BotActionFlags.GrenadeActive)
 					p_Bot._ShootModeTimer = p_Bot._ShootModeTimer + 2 * Registry.BOT.BOT_UPDATE_CYCLE
 				end
@@ -133,7 +132,7 @@ function BotAttacking:UpdateAttacking(p_Bot)
 					end
 					-- use grenade from time to time
 					if Config.BotsThrowGrenades then
-						local s_TargetTimeValue = Config.BotFireModeDuration - 2 * Registry.BOT.BOT_UPDATE_CYCLE
+						local s_TargetTimeValue = Config.BotFireModeDuration - (3 * Registry.BOT.BOT_UPDATE_CYCLE)
 
 						if ((p_Bot._ShootModeTimer >= (s_TargetTimeValue - 0.001)) and
 							(p_Bot._ShootModeTimer <= (s_TargetTimeValue + Registry.BOT.BOT_UPDATE_CYCLE + 0.001)) and
@@ -176,7 +175,16 @@ function BotAttacking:UpdateAttacking(p_Bot)
 			end
 
 			--shooting sequence
-			if p_Bot.m_ActiveWeapon ~= nil then
+			if Globals.IsGm then
+				if p_Bot._ShotTimer >= (0.4) then
+					p_Bot._ShotTimer = 0.0
+				end
+				if p_Bot._ShotTimer <= 0.2 then
+					p_Bot:_SetInput(EntryInputActionEnum.EIAFire, 1)
+				end
+
+				p_Bot._ShotTimer = p_Bot._ShotTimer + Registry.BOT.BOT_UPDATE_CYCLE
+			elseif p_Bot.m_ActiveWeapon ~= nil then
 				if p_Bot.m_KnifeMode then
 					-- nothing to do
 					-- C4 Handling