+++ .github/CHANGELOG.md
@@ -3,8 +3,8 @@
 ![Image](https://img.shields.io/github/downloads/Joe91/fun-bots/total?style=for-the-badge)
 ![Image](https://img.shields.io/github/stars/Joe91/fun-bots?style=for-the-badge)
 
-## Welcome to the changelogs for release **V2.6.0**
-This is the changelog for the version V2.6.0. Don't forget to [join us on Discord](https://discord.funbots.dev)
+## Welcome to the changelogs for release **V2.6.1**
+This is the changelog for the version V2.6.1. Don't forget to [join us on Discord](https://discord.funbots.dev)
 
 ## Changelog
 
@@ -27,6 +27,15 @@ This is the changelog for the version V2.6.0. Don't forget to [join us on Discor
 * massive rework on NodeEditor: Now completely on Server
 * count paths correctly
 * delte existing path now possible
+* Bots use flares or smoke
+
+* Limit Distance of Bots when they shoot back
+* Use Statevents to increase Rush-Stability
+* Bots use Squad-Perks (#101)
+* First implementation of Alternate Weapons on some vehicles (#172)
+* Spawn at Player, if player close to path
+* Don't use missiles without vehicles
+* Don't spam nades on vehicles anymore
 
 ### Bug fixes
 * #218 increased auto-sniper-time a little
@@ -37,8 +46,12 @@ This is the changelog for the version V2.6.0. Don't forget to [join us on Discor
 * modified mapfiles with newlines can be imported now
 * several bugs in new node-editor
 
+* fix endless attack of vehicles
+* fix Gunmaster Gamemode
+* fix Bots keeping the knife for one whole fight.
+
 ### New maps
-* 
+* Rush Wake Island
 
 ### Updated maps
 * XP4_Parl_ConquestLarge0 (fixed some paths #219)